Manual Student Assigner
A work in progress...some frontend changes have been made
